consolidate fs.go, internal/fs.go and gps/internal/fs/fs/.go into internal/fs
