+++ context.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 
 	"github.com/Masterminds/vcs"
-	"github.com/golang/dep/internal"
+	"github.com/golang/dep/internal/fs"
 	"github.com/golang/dep/internal/gps"
 	"github.com/pkg/errors"
 )
@@ -35,7 +35,7 @@ func NewContext(wd string, env []string) (*Ctx, error) {
 	for _, gp := range filepath.SplitList(GOPATH) {
 		gp = filepath.FromSlash(gp)
 
-		if internal.HasFilepathPrefix(filepath.FromSlash(wd), gp) {
+		if fs.HasFilepathPrefix(filepath.FromSlash(wd), gp) {
 			ctx.GOPATH = gp
 		}
 
@@ -198,7 +198,7 @@ func (c *Ctx) resolveProjectRoot(path string) (string, error) {
 	// Determine if the symlink is within any of the GOPATHs, in which case we're not
 	// sure how to resolve it.
 	for _, gp := range c.GOPATHS {
-		if internal.HasFilepathPrefix(path, gp) {
+		if fs.HasFilepathPrefix(path, gp) {
 			return "", errors.Errorf("'%s' is linked to another path within a GOPATH (%s)", path, gp)
 		}
 	}
@@ -213,7 +213,7 @@ func (c *Ctx) resolveProjectRoot(path string) (string, error) {
 // The second returned string indicates which GOPATH value was used.
 func (c *Ctx) SplitAbsoluteProjectRoot(path string) (string, error) {
 	srcprefix := filepath.Join(c.GOPATH, "src") + string(filepath.Separator)
-	if internal.HasFilepathPrefix(path, srcprefix) {
+	if fs.HasFilepathPrefix(path, srcprefix) {
 		// filepath.ToSlash because we're dealing with an import path now,
 		// not an fs path
 		return filepath.ToSlash(path[len(srcprefix):]), nil
@@ -227,7 +227,7 @@ func (c *Ctx) SplitAbsoluteProjectRoot(path string) (string, error) {
 // package directory needs to exist.
 func (c *Ctx) absoluteProjectRoot(path string) (string, error) {
 	posspath := filepath.Join(c.GOPATH, "src", path)
-	dirOK, err := IsDir(posspath)
+	dirOK, err := fs.IsDir(posspath)
 	if err != nil {
 		return "", errors.Wrapf(err, "checking if %s is a directory", posspath)
 	}
