--- internal/gps/internal/fs/fs_test.go
@@ -1,135 +1,0 @@
-// Copyright 2016 The Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package fs
-
-import (
-	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"testing"
-)
-
-func isDir(name string) (bool, error) {
-	fi, err := os.Stat(name)
-	if os.IsNotExist(err) {
-		return false, nil
-	}
-	if err != nil {
-		return false, err
-	}
-	if !fi.IsDir() {
-		return false, fmt.Errorf("%q is not a directory", name)
-	}
-	return true, nil
-}
-
-func TestCopyDir(t *testing.T) {
-	dir, err := ioutil.TempDir("", "gps")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	srcdir := filepath.Join(dir, "src")
-	if err := os.MkdirAll(srcdir, 0755); err != nil {
-		t.Fatal(err)
-	}
-
-	srcf, err := os.Create(filepath.Join(srcdir, "myfile"))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	contents := "hello world"
-	if _, err := srcf.Write([]byte(contents)); err != nil {
-		t.Fatal(err)
-	}
-	srcf.Close()
-
-	destdir := filepath.Join(dir, "dest")
-	if err := CopyDir(srcdir, destdir); err != nil {
-		t.Fatal(err)
-	}
-
-	dirOK, err := isDir(destdir)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !dirOK {
-		t.Fatalf("expected %s to be a directory", destdir)
-	}
-
-	destf := filepath.Join(destdir, "myfile")
-	destcontents, err := ioutil.ReadFile(destf)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if contents != string(destcontents) {
-		t.Fatalf("expected: %s, got: %s", contents, string(destcontents))
-	}
-
-	srcinfo, err := os.Stat(srcf.Name())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	destinfo, err := os.Stat(destf)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if srcinfo.Mode() != destinfo.Mode() {
-		t.Fatalf("expected %s: %#v\n to be the same mode as %s: %#v", srcf.Name(), srcinfo.Mode(), destf, destinfo.Mode())
-	}
-}
-
-func TestCopyFile(t *testing.T) {
-	dir, err := ioutil.TempDir("", "gps")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	srcf, err := os.Create(filepath.Join(dir, "srcfile"))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	contents := "hello world"
-	if _, err := srcf.Write([]byte(contents)); err != nil {
-		t.Fatal(err)
-	}
-	srcf.Close()
-
-	destf := filepath.Join(dir, "destf")
-	if err := CopyFile(srcf.Name(), destf); err != nil {
-		t.Fatal(err)
-	}
-
-	destcontents, err := ioutil.ReadFile(destf)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if contents != string(destcontents) {
-		t.Fatalf("expected: %s, got: %s", contents, string(destcontents))
-	}
-
-	srcinfo, err := os.Stat(srcf.Name())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	destinfo, err := os.Stat(destf)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if srcinfo.Mode() != destinfo.Mode() {
-		t.Fatalf("expected %s: %#v\n to be the same mode as %s: %#v", srcf.Name(), srcinfo.Mode(), destf, destinfo.Mode())
-	}
-}
